+++ src/app/components/TokensBottomBar.tsx
@@ -1,42 +1,25 @@
 import React, { useCallback } from 'react';
 import { useSelector } from 'react-redux';
 import ApplySelector from './ApplySelector';
-import ExportModal from './modals/ExportModal';
-import PresetModal from './modals/PresetModal';
 import Box from './Box';
 import StylesDropdown from './StylesDropdown';
-import { editProhibitedSelector, hasUnsavedChangesSelector } from '@/selectors';
+import { hasUnsavedChangesSelector } from '@/selectors';
 import Button from './Button';
 import Stack from './Stack';
 import SettingsDropdown from './SettingsDropdown';
 import useTokens from '../store/useTokens';
 import { stringTokensSelector } from '@/selectors/stringTokensSelector';
+import ToolsDropdown from './ToolsDropdown';
 
 type Props = {
   hasJSONError: boolean;
 };
 
 export default function TokensBottomBar({ hasJSONError }: Props) {
-  const editProhibited = useSelector(editProhibitedSelector);
   const hasUnsavedChanges = useSelector(hasUnsavedChangesSelector);
   const stringTokens = useSelector(stringTokensSelector);
 
   const { handleJSONUpdate } = useTokens();
-
-  const [exportModalVisible, showExportModal] = React.useState(false);
-  const [presetModalVisible, showPresetModal] = React.useState(false);
-
-  const handleShowPresetModal = useCallback(() => {
-    showPresetModal(true);
-  }, []);
-
-  const handleClosePresetModal = useCallback(() => {
-    showPresetModal(false);
-  }, []);
-
-  const handleCloseExportModal = useCallback(() => {
-    showExportModal(false);
-  }, []);
 
   const handleSaveJSON = useCallback(() => {
     handleJSONUpdate(stringTokens);
@@ -75,9 +58,7 @@ export default function TokensBottomBar({ hasJSONError }: Props) {
             }}
           >
             <Stack direction="row" gap={1}>
-              <Button variant="ghost" disabled={editProhibited} onClick={handleShowPresetModal}>
-                Tools
-              </Button>
+              <ToolsDropdown />
               <StylesDropdown />
             </Stack>
             <Stack direction="row" gap={1}>
@@ -86,8 +67,6 @@ export default function TokensBottomBar({ hasJSONError }: Props) {
             </Stack>
           </Stack>
         )}
-      {exportModalVisible && <ExportModal onClose={handleCloseExportModal} />}
-      {presetModalVisible && <PresetModal onClose={handleClosePresetModal} />}
     </Box>
   );
 }
